+++ src/ref.ts
@@ -1,6 +1,6 @@
 import { isObject } from "./util"
 
-interface Ref<T extends Node> {
+interface Ref<T = Node> {
   current: null | T
 }
 
@@ -8,6 +8,6 @@ export function createRef<T extends Node = Node>(): Ref<T> {
   return Object.seal({ current: null })
 }
 
-export function isRef<T extends Node>(maybeRef: any): maybeRef is Ref<T> {
+export function isRef<T = Node>(maybeRef: any): maybeRef is Ref<T> {
   return isObject(maybeRef) && "current" in maybeRef
 }